Many fixes after reunion
* KronMatrix:
- Change function name: splitA -> ReplicateA

* MultMatrix:
- Improve matrix multiplication constraint verification
- Remove temp directory from the input

* QW:
- Change MultMatrix call

* QWS:
- Change MultMatrix call
- Change the name for Quandoop
- Change configurations name and definition
- newShape -> dimensions
- axes -> measurement
- Removed the requirement to set the configurations to calculate the
probability distribution

+++ Quandoop/src/Quandoop.java
@@ -1,5 +1,5 @@
 /*
- * QWS.java    1.0 2015/05/20
+ * Quandoop.java    1.2 2015/07/13
  *
  * Copyright (C) 2015 GNU General Public License
  *
@@ -15,7 +15,7 @@
  */
 
 
-package qws;
+package quandoop;
 
 import java.io.File;
 import java.io.FileReader;
@@ -39,11 +39,11 @@ import org.apache.hadoop.conf.Configuration;
     with a sequence of matrices multiplication using Apache Hadoop.
  *
  * @version
-    1.0 20 May 2015  * @author
+    1.2 13 Jul 2015  * @author
     David Souza  */
 
 
-public class QWS {
+public class Quandoop {
 
     public static void main(String[] args) throws Exception {
 
@@ -52,8 +52,8 @@ public class QWS {
         String paths;
         String workDir;
         String jarDir;
-        String newShape;
-        String axes;
+        String dimensions;
+        String measurement;
         String position;
         String outputDir;
         String line;
@@ -62,7 +62,10 @@ public class QWS {
         String psiTNorm;
         String absSquare;
         String reshape;
-        String pdf;
+        String pdf = "";
+        boolean pdfCalc = false;
+        int numAxes;
+        String fullAxes;
         BufferedReader br;
         int numberU;
         String[] uDir;
@@ -93,37 +96,63 @@ public class QWS {
 		    paths = prop.getProperty("paths");
 		    workDir = prop.getProperty("workDir");
 		    jarDir = prop.getProperty("jarDir");
-		    newShape = prop.getProperty("newShape");
-		    axes = prop.getProperty("axes");
+		    dimensions = prop.getProperty("dimensions");
+		    measurement = prop.getProperty("measurement");
 		    position = prop.getProperty("position");
 		    outputDir = prop.getProperty("outputDir");
 
             if (steps == null || steps.equals("")) {
-                throw new IOException ("The value of the configuration "
+                throw new IOException("The value of the configuration "
                         + "\"steps\" can not be null or empty.");
             }
 
             if (paths == null || paths.equals("")) {
-                throw new IOException ("The value of the configuration "
+                throw new IOException("The value of the configuration "
                         + "\"paths\" can not be null or empty.");
             } 
 
             if (workDir == null || workDir.equals("")) {
-                throw new IOException ("The value of the configuration "
+                throw new IOException("The value of the configuration "
                         + "\"workDir\" can not be null or empty.");
             }
 
             if (jarDir == null || jarDir.equals("")) {
-                throw new IOException ("The value of the configuration "
+                throw new IOException("The value of the configuration "
                         + "\"jarDir\" can not be null or empty.");
             }
 
             if (outputDir == null || outputDir.equals("")) {
-                throw new IOException ("The value of the configuration "
+                throw new IOException("The value of the configuration "
                         + "\"outputDir\" can not be null or empty.");
             }
 
+            if (dimensions == null || dimensions.equals("")
+		            || measurement == null || measurement.equals("")
+		            || position  == null || position.equals("")) {
+
+                pdfCalc = false;
+                System.out.println("The value of the configuration "
+                        + "\"dimensions\", \"measurement\" or \"position\" is "
+                        + "null or empty. The pdf calculation will not be run."
+                        );
+
+            } else {
+                pdfCalc = true;
+            }
+
             configInput.close();
+
+            if (pdfCalc) {
+
+                numAxes = dimensions.split(",").length - position.
+                            replaceAll("/?,", "").split(",").length;
+                fullAxes = "1";
+                for (int i = 2; i < numAxes; i++) {
+                    fullAxes += "," + Integer.toString(i);
+                }
+                measurement = fullAxes.replaceAll(measurement + ",", "");
+
+            }
 
             /*
              * Delete the workDir directory if exists. And create a new one
@@ -257,8 +286,7 @@ public class QWS {
 
                     pr = rt.exec("hadoop jar " + jarDir
                             + "operations.jar operations.MultMatrix " + psiT
-                            + "_tmp" + " " + workDir + "tmp" + " " + psiT
-                            + " B");
+                            + "_tmp" + " " + psiT + " B");
 
                     pr.waitFor();
 
@@ -334,108 +362,112 @@ public class QWS {
                     null);
             fs.rename(new Path(psiT + "_New"), pt);
 
-            // Start of the absSquare
-            absSquare = workDir + "absSquare";
+            if (pdfCalc) {
 
-            // Delete the output directory if exists.
-            pt = new Path(absSquare);
-            fs.delete(pt, true);
+                // Start of the absSquare
+                absSquare = workDir + "absSquare";
 
-            /*
-             * Computes the square of the absolute value for each element of the
-             * array.
-             */
-            pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
-                    + "AbsSquare " + psiT + " " + absSquare);
+                // Delete the output directory if exists.
+                pt = new Path(absSquare);
+                fs.delete(pt, true);
 
-            pr.waitFor();
+                /*
+                 * Computes the square of the absolute value for each element of the
+                 * array.
+                 */
+                pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
+                        + "AbsSquare " + psiT + " " + absSquare);
 
-            br = new BufferedReader(new InputStreamReader(
-                    pr.getInputStream()));
+                pr.waitFor();
 
-            if (pr.exitValue() != 0) {
+                br = new BufferedReader(new InputStreamReader(
+                        pr.getInputStream()));
 
-                while ((line = br.readLine()) != null) {
-                    System.out.println(line);
+                if (pr.exitValue() != 0) {
+
+                    while ((line = br.readLine()) != null) {
+                        System.out.println(line);
+                    }
+
+                    System.exit(1);
                 }
 
-                System.exit(1);
-            }
+                pr.destroy();
 
-            pr.destroy();
+                // End of the absSquare
 
-            // End of the absSquare
+                // Start of the reshape
+                reshape = workDir + "reshape";
 
-            // Start of the reshape
-            reshape = workDir + "reshape";
+                // Delete the output directory if exists.
+                pt = new Path(reshape);
+                fs.delete(pt, true);
 
-            // Delete the output directory if exists.
-            pt = new Path(reshape);
-            fs.delete(pt, true);
+                // Gives a new shape for the array.
+                pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
+                        + "Reshape " + dimensions + " " + absSquare + " "
+                        + reshape);
 
-            // Gives a new shape for the array.
-            pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
-                    + "Reshape " + newShape + " " + absSquare + " " + reshape);
+                pr.waitFor();
 
-            pr.waitFor();
+                br = new BufferedReader(new InputStreamReader(
+                        pr.getInputStream()));
 
-            br = new BufferedReader(new InputStreamReader(
-                    pr.getInputStream()));
+                if (pr.exitValue() != 0) {
 
-            if (pr.exitValue() != 0) {
+                    while ((line = br.readLine()) != null) {
+                        System.out.println(line);
+                    }
 
-                while ((line = br.readLine()) != null) {
-                    System.out.println(line);
+                    System.exit(1);
                 }
 
-                System.exit(1);
-            }
+                pr.destroy();
 
-            pr.destroy();
+                // End of the reshape
 
-            // End of the reshape
+                /*
+                 * Start of the sumAxis. In this case the output of SunAxis function
+                 * will be the PDF of psiT.
+                 */
+                pdf = workDir + "pdf";
 
-            /*
-             * Start of the sumAxis. In this case the output of SunAxis function
-             * will be the PDF of psiT.
-             */
-            pdf = workDir + "pdf";
+                // Delete the output directory if exists.
+                pt = new Path(pdf);
+                fs.delete(pt, true);
 
-            // Delete the output directory if exists.
-            pt = new Path(pdf);
-            fs.delete(pt, true);
+                /*
+                 * Sum the elements of the array over given measurement, for a
+                 * specific position.
+                 */ 
+                pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
+                        + "SumAxis " + measurement + " " +  position + " " + reshape
+                        + " " + pdf);
 
-            /*
-             * Sum the elements of the array over given axes, for a specific
-             * position.
-             */ 
-            pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
-                    + "SumAxis " + axes + " " +  position + " " + reshape + " "
-                    + pdf);
+                pr.waitFor();
 
-            pr.waitFor();
+                br = new BufferedReader(new InputStreamReader(
+                        pr.getInputStream()));
 
-            br = new BufferedReader(new InputStreamReader(
-                    pr.getInputStream()));
+                if (pr.exitValue() != 0) {
 
-            if (pr.exitValue() != 0) {
+                    while ((line = br.readLine()) != null) {
+                        System.out.println(line);
+                    }
 
-                while ((line = br.readLine()) != null) {
-                    System.out.println(line);
+                    System.exit(1);
                 }
 
-                System.exit(1);
+                pr.destroy();
+
+                // End of the sumAxis
+
+                System.out.println("End of the pdf.");
             }
 
-            pr.destroy();
-
-            // End of the sumAxis
-
-            System.out.println("End of the pdf.");
-
             /*
-             * Delete _logs folder and _SUCCESS file in the psiT and psiTNorm
-             * folders.
+             * Delete _logs folder and _SUCCESS file in the psiT, psiTNorm and
+             * pdf folders.
              */
             pt = new Path(psiT + "/_logs");
             fs.delete(pt, true);
@@ -445,10 +477,12 @@ public class QWS {
             fs.delete(pt, true);
             pt = new Path(psiTNorm + "/_SUCCESS");
             fs.delete(pt, false);
-            pt = new Path(pdf + "/_logs");
-            fs.delete(pt, true);
-            pt = new Path(pdf + "/_SUCCESS");
-            fs.delete(pt, false);
+            if (pdfCalc) {
+                pt = new Path(pdf + "/_logs");
+                fs.delete(pt, true);
+                pt = new Path(pdf + "/_SUCCESS");
+                fs.delete(pt, false);
+            }
 
             // Delete the outputDir if it exists.
             fu.fullyDelete(new File(outputDir));
@@ -458,8 +492,10 @@ public class QWS {
             fu.copy(fs, pt, new File(outputDir + "psiT"), false, conf);
             pt = new Path(psiTNorm);
             fu.copy(fs, pt, new File(outputDir + "psiTNorm"), false, conf);
-            pt = new Path(pdf);
-            fu.copy(fs, pt, new File(outputDir + "pdf"), false, conf);
+            if (pdfCalc) {
+                pt = new Path(pdf);
+                fu.copy(fs, pt, new File(outputDir + "pdf"), false, conf);
+            }
 
             // Delete the workDir directory.
             pt = new Path(workDir);
